allez hop
l'argument de syscall (caller_tss_selector) est pas utilisé et c'est
surement la cause du fait que RIEN NE MARCHE. Voila

+++ groupe13/OS/user/ulibc.c
@@ -162,8 +162,8 @@ void sleep(uint ms) {
 	
 }
 
-int get_ticks() { 
-	return syscall(SYSCALL_GET_TICKS, 0, 0, 0, 0, 0);
+uint get_ticks() { 
+	return (uint) syscall(SYSCALL_GET_TICKS, 0, 0, 0, 0, 0);
 }
 
 char *
